+++ src/example/index.ts
@@ -1,5 +1,6 @@
-import { GEView } from "../index";
-import { GEShapeTypes, GEShapeName, GENode, GEEdge } from "../types";
+import { GEView, GENode, GEEdge } from "../index";
+import { nodeTypes, edgeTypes } from "./node-types";
+import { randomize } from "./utils";
 
 const graphDiv = document.getElementById("graph");
 const nodeCountSpan = document.getElementById("node-count-span");
@@ -9,81 +10,6 @@ const generateTextbox = document.getElementById(
   "generate-textbox"
 ) as HTMLInputElement;
 const generateButton = document.getElementById("generate-button");
-
-const nodeTypes: GEShapeTypes = {
-  empty: [
-    {
-      shape: GEShapeName.CIRCLE,
-      r: 80
-    }
-  ],
-  decision: [
-    {
-      shape: GEShapeName.RECTANGLE,
-      width: 150,
-      height: 120
-    }
-  ],
-  unknown: [
-    {
-      shape: GEShapeName.POLYGON,
-      points: [
-        [0, -80],
-        [80, 0],
-        [0, 80],
-        [-80, 0]
-      ]
-    }
-  ],
-  complex: [
-    {
-      shape: GEShapeName.CIRCLE,
-      r: 80
-    },
-    {
-      shape: GEShapeName.CIRCLE,
-      r: 60,
-      color: "#9AE6B4"
-    },
-    {
-      shape: GEShapeName.CIRCLE,
-      r: 40,
-      color: "white"
-    }
-  ]
-};
-
-const edgeTypes: GEShapeTypes = {
-  normal: [
-    {
-      shape: GEShapeName.POLYGON,
-      points: [
-        [0, -25],
-        [25, 0],
-        [0, 25],
-        [-25, 0]
-      ]
-    }
-  ],
-  round: [
-    {
-      shape: GEShapeName.CIRCLE,
-      r: 25
-    }
-  ],
-  double: [
-    {
-      shape: GEShapeName.CIRCLE,
-      r: 25,
-      color: "#E9D8FD"
-    },
-    {
-      shape: GEShapeName.CIRCLE,
-      r: 15,
-      color: "white"
-    }
-  ]
-};
 
 let nodes: GENode[] = [];
 let edges: GEEdge[] = [];
@@ -156,76 +82,22 @@ function handleMoveNode(node: GENode, newX: number, newY: number) {
   node.y = newY;
 }
 
+function handleViewZoom() {
+  zoomSlider.value = graphView.getScale().toString();
+}
+
 graphView.setOptions({
   minScale: 0.2,
   maxScale: 3.0,
   nodeTypes,
   edgeTypes,
-  onViewZoom: () => {
-    zoomSlider.value = graphView.getScale().toString();
-  },
+  onViewZoom: handleViewZoom,
   onCreateNode: handleCreateNode,
   onCreateEdge: handleCreateEdge,
   onDeleteNode: handleDeleteNode,
   onDeleteEdge: handleDeleteEdge,
   onMoveNode: handleMoveNode
 });
-
-function getRandomIntInclusive(minF: number, maxF: number): number {
-  const min = Math.ceil(minF);
-  const max = Math.floor(maxF);
-  return Math.floor(Math.random() * (max - min + 1) + min); //The maximum is inclusive and the minimum is inclusive
-}
-
-function randomize(nodeCount = 1000, cols = 40) {
-  nodes = [];
-  edges = [];
-  lastId = 0;
-
-  for (let i = 0; i < nodeCount; i++) {
-    const col = i % cols;
-    const row = Math.floor(i / cols);
-
-    const tmp = getRandomIntInclusive(0, 3);
-    const nodeType =
-      tmp === 0
-        ? "empty"
-        : tmp === 1
-        ? "decision"
-        : tmp === 2
-        ? "unknown"
-        : "complex";
-
-    const tmp2 = getRandomIntInclusive(0, 2);
-    const edgeType = tmp2 === 0 ? "normal" : tmp2 === 1 ? "round" : "double";
-
-    lastId++;
-    const currNode: GENode = {
-      id: lastId,
-      x: col * 320,
-      y: row * 320,
-      type: nodeType,
-      text: `Node ID: ${lastId}`
-    };
-
-    nodes.push(currNode);
-
-    if (i > 0) {
-      const prevNode = nodes[i - 1];
-
-      lastId++;
-      edges.push({
-        id: lastId,
-        sourceNode: prevNode,
-        targetNode: currNode,
-        type: edgeType,
-        text: lastId.toString()
-      });
-    }
-  }
-
-  graphView.setData(nodes, edges);
-}
 
 graphView.init(graphDiv);
 
@@ -243,7 +115,13 @@ generateButton.addEventListener("click", () => {
   const value = parseInt(generateTextbox.value, 10);
   const columns = Math.ceil(Math.sqrt(value));
 
-  randomize(value, columns);
+  const r = randomize(value, columns);
+
+  lastId = r.lastId;
+  nodes = r.nodes;
+  edges = r.edges;
+
+  graphView.setData(r.nodes, r.edges);
 
   updateNodeCount();
   updateEdgeCount();
